+++ tests/integration/cases/record_named_shape/Dart@v3.dart
@@ -1,0 +1,4 @@
+final my_data = <Map<String, dynamic>>[
+    <String, dynamic>{"id": 100, "description": "first task", "is_done": false, "blocks": <int>[102, 103]},
+    <String, dynamic>{"id": 101, "description": "second task", "is_done": true, "blocks": <int>[100]},
+];