+++ tests/integration/cases/record_named_shape/JavaScript_combined@es2015.js
@@ -1,0 +1,8 @@
+let my_data = [
+  {"id": 100, "description": "first task", "is_done": false, "blocks": [102, 103]},
+  {"id": 101, "description": "second task", "is_done": true, "blocks": [100]},
+];
+my_data = [
+  {"id": 100, "description": "first task", "is_done": false, "blocks": [102, 103]},
+  {"id": 101, "description": "second task", "is_done": true, "blocks": [100]},
+];